requirements.txt
# Core dependencies
numpy==1.24.3
opencv-python==4.8.0.76
ffmpeg-python==0.2.0
python-dotenv==1.0.0

# NLP and ML
google-generativeai==0.3.1

# OCR
pytesseract==0.3.10

# Web and templating
jinja2==3.1.2
markdown==3.4.4

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Development
black==23.7.0
flake8==6.1.0
mypy==1.5.1
isort==5.12.0